fix(nextjs): remove unecessary prop
