+++ apps/nextjs/src/components/nav/email-account-nav.tsx
@@ -2,46 +2,62 @@
 
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { Avatar, AvatarFallback } from "@radix-ui/react-avatar";
+
+import { useAllEmailProviders } from "@skylar/client-db";
 
 import { cn } from "~/lib/ui";
-import { AvatarImage } from "../ui/avatar";
-
-type SidebarNavProps = {
-  allEmailProvidersProfileInfo: {
-    href: string;
-    title: string;
-    imageUri: string;
-    name: string;
-  }[];
-} & React.HTMLAttributes<HTMLElement>;
+import { Avatar, AvatarFallback, AvatarImage } from "../ui/avatar";
+import { Skeleton } from "../ui/skeleton";
 
 export function EmailAccountNav({
   className,
-  allEmailProvidersProfileInfo,
   ...props
-}: SidebarNavProps) {
+}: React.HTMLAttributes<HTMLElement>) {
   const pathname = usePathname();
+  const { data: allEmailProviders, isLoading } = useAllEmailProviders();
 
-  console.log("allEmailProvidersProfileInfo", allEmailProvidersProfileInfo);
+  const allEmailProvidersProfileInfo = (allEmailProviders ?? []).map(
+    (account) => ({
+      href: `${account.provider_id}`,
+      title: account.email,
+      imageUri: account.image_uri,
+      name: account.inbox_name,
+    }),
+  );
+
+  let body = (
+    <>
+      <Skeleton className="h-10 w-10 rounded-full" />
+      <Skeleton className="h-10 w-10 rounded-full" />
+      <Skeleton className="h-10 w-10 rounded-full" />
+    </>
+  );
+  if (!isLoading) {
+    body = (
+      <>
+        {allEmailProvidersProfileInfo.map((item) => (
+          <Link
+            key={item.href}
+            href={item.href}
+            className={cn(
+              pathname === item.href
+                ? "hover:cursor-default"
+                : "transition-opacity hover:cursor-pointer hover:opacity-70",
+            )}
+          >
+            <Avatar>
+              <AvatarImage src={item.imageUri} />
+              <AvatarFallback>{item.name.slice(0, 2)}</AvatarFallback>
+            </Avatar>
+          </Link>
+        ))}
+      </>
+    );
+  }
+
   return (
     <nav className={cn("flex  flex-col gap-5", className)} {...props}>
-      {allEmailProvidersProfileInfo.map((item) => (
-        <Link
-          key={item.href}
-          href={item.href}
-          className={cn(
-            pathname === item.href
-              ? "hover:cursor-default"
-              : "transition-opacity hover:cursor-pointer hover:opacity-70",
-          )}
-        >
-          <Avatar>
-            <AvatarImage src={item.imageUri} />
-            <AvatarFallback>{item.name.slice(0, 2)}</AvatarFallback>
-          </Avatar>
-        </Link>
-      ))}
+      {body}
     </nav>
   );
 }
